+++ samples/HelloTriangle/main.cpp
@@ -339,7 +339,7 @@ public:
 
         vk::CommandBufferAllocateInfo allocInfo = {};
         allocInfo.commandPool = commandPool.get();
-        allocInfo.commandBufferCount = static_cast<uint32_t>(swapchain->images().size());
+        allocInfo.commandBufferCount = swapchain->images().size();
         allocInfo.level = vk::CommandBufferLevel::Primary;
 
         commandBuffers = commandPool->allocate(allocInfo);
